move spdm_transport_none_lib to a lib.
