+++ library/spdm_transport_none_lib/CMakeLists.txt
@@ -1,0 +1,13 @@
+cmake_minimum_required(VERSION 2.6)
+
+INCLUDE_DIRECTORIES(${LIBSPDM_DIR}/include
+                    ${LIBSPDM_DIR}/include/hal/${ARCH}
+                    ${PROJECT_SOURCE_DIR}/include
+)
+
+SET(src_spdm_transport_none_lib
+    common.c
+    none.c
+)
+
+ADD_LIBRARY(spdm_transport_none_lib STATIC ${src_spdm_transport_none_lib})
