+++ library/spdm_transport_none_lib/common.c
@@ -4,7 +4,7 @@
     License: BSD 3-Clause License. For full text see link: https://github.com/DMTF/libspdm/blob/main/LICENSE.md
 **/
 
-#include "spdm_transport_none_lib.h"
+#include "library/spdm_transport_none_lib.h"
 #include "library/spdm_secured_message_lib.h"
 
 /**
