+++ spdm_emu/spdm_requester_emu/CMakeLists.txt
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 2.6)
 
 INCLUDE_DIRECTORIES(${PROJECT_SOURCE_DIR}/spdm_emu/spdm_requester_emu
                     ${PROJECT_SOURCE_DIR}/spdm_emu/spdm_emu_common
-                    ${PROJECT_SOURCE_DIR}/spdm_emu/spdm_transport_none_lib
+                    ${PROJECT_SOURCE_DIR}/include
                     ${LIBSPDM_DIR}/os_stub/spdm_device_secret_lib_sample
                     ${LIBSPDM_DIR}/include
                     ${LIBSPDM_DIR}/include/hal/${ARCH}
@@ -21,8 +21,6 @@ SET(src_spdm_requester_emu
     ${PROJECT_SOURCE_DIR}/spdm_emu/spdm_emu_common/nv_storage.c
     ${PROJECT_SOURCE_DIR}/spdm_emu/spdm_emu_common/pcap.c
     ${PROJECT_SOURCE_DIR}/spdm_emu/spdm_emu_common/support.c
-    ${PROJECT_SOURCE_DIR}/spdm_emu/spdm_transport_none_lib/common.c
-    ${PROJECT_SOURCE_DIR}/spdm_emu/spdm_transport_none_lib/none.c
 )
 
 SET(spdm_requester_emu_LIBRARY
@@ -38,6 +36,7 @@ SET(spdm_requester_emu_LIBRARY
     spdm_secured_message_lib
     spdm_transport_mctp_lib
     spdm_transport_pcidoe_lib
+    spdm_transport_none_lib
     spdm_device_secret_lib_sample
 )
