+++ spdm_emu/spdm_requester_emu/spdm_requester_emu.h
@@ -10,7 +10,7 @@
 #include "hal/base.h"
 #include "hal/library/memlib.h"
 #include "library/spdm_requester_lib.h"
-#include "spdm_transport_none_lib.h"
+#include "library/spdm_transport_none_lib.h"
 #include "library/spdm_transport_mctp_lib.h"
 #include "library/spdm_transport_pcidoe_lib.h"
