+++ spdm_emu/spdm_responder_emu/spdm_responder_emu.h
@@ -10,7 +10,7 @@
 #include "hal/base.h"
 #include "hal/library/memlib.h"
 #include "library/spdm_responder_lib.h"
-#include "spdm_transport_none_lib.h"
+#include "library/spdm_transport_none_lib.h"
 #include "library/spdm_transport_mctp_lib.h"
 #include "library/spdm_transport_pcidoe_lib.h"
